requirements.txt
streamlit>=1.32.0
langchain-core>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.1
python-dotenv>=1.0.0